added landing page,navigation bar and routes
