+++ src/Components/AboutUs/AboutUs.js
@@ -1,0 +1,12 @@
+import NavigationBar from "../NavgationBar/NavigationBar";
+
+function AboutUs(){
+    return(
+        <>
+        <NavigationBar/>
+        <h1> About Us page - development in progress</h1>
+        </>
+    )
+}
+
+export default AboutUs;
