+++ src/Components/ContactUs/ContactUs.js
@@ -1,0 +1,12 @@
+import NavigationBar from "../NavgationBar/NavigationBar";
+
+function ContactUs(){
+    return(
+        <>
+        <NavigationBar/>
+        <h1> ContactUs page - development in progress</h1>
+        </>
+    )
+}
+
+export default ContactUs;
