+++ src/Components/NoRoute.js
@@ -1,0 +1,9 @@
+function NoMatch() {
+    return (
+      <div style={{ padding: 20 }}>
+        <h2>404: Page Not Found</h2>
+        <p>Lorem ipsum dolor sit amet, consectetur adip.</p>
+      </div>
+    );
+  }
+  export default NoMatch;
