--- src/Components/NurseDashboard.js
@@ -1,9 +1,0 @@
-
-function NurseDashboard(){
-
-    return (
-        <div>
-            
-        </div>
-    )
-}
