+++ src/Components/SignupAndLogin/OptionList.js
@@ -1,0 +1,21 @@
+const roleList = [
+    { value: "Doctor", label: "Doctor" },
+    { value: "Nurse", label: "Nurse" },
+    { value: "Receptionist", label: "Receptionist" },
+  ];
+
+  const countryList = [{ value: "India", label: "India" }];
+
+  const stateList = [
+    { value: "Delhi", label: "Delhi" },
+    { value: "Karnataka", label: "Karnataka" },
+    { value: "Tamil Nadu", label: "Tamil Nadu" },
+  ];
+
+  const townCityList = [
+    { value: "Delhi", label: "Delhi" },
+    { value: "Bangalore", label: "Bangalore" },
+    { value: "Chennai", label: "Chennai" },
+  ];
+
+  export {roleList,countryList,stateList,townCityList}
