+++ src/Components/UserDashboards/Nurse/NurseDashboard.js
@@ -1,0 +1,9 @@
+
+function NurseDashboard(){
+
+    return (
+        <div>
+            
+        </div>
+    )
+}
