+++ src/index.js
@@ -1,15 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import Signup from './Components/Signup';
-import Login from './Components/Login';
 import reportWebVitals from './reportWebVitals';
+import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <Login/>      
-    <Signup/>
+    <App/>
   </React.StrictMode>
 );
